Added Mass-Rolls
Mass Rolls allows you to, as a DM or player, deal with those situations where you need to make a very large number of attack rolls or saves for hordes of creatures. More complex versions exist with more features but this one is a bare bones version that's quick and easy to implement and use.